+++ R/fst.R
@@ -27,7 +27,7 @@
 #'
 #' # estimate the kinship matrix "Phi" from the genotypes "X"!
 #' Phi <- popkin(X, subpops) # calculate kinship from X and optional subpop labels
-#' w <- weightsSubpops(subpops) # can weigh individuals so subpopulations are balanced
+#' w <- weights_subpops(subpops) # can weigh individuals so subpopulations are balanced
 #' Fst <- fst(Phi, w) # use kinship matrix and weights to calculate fst
 #' 
 #' Fst <- fst(Phi) # no (or NULL) weights implies uniform weights