+++ R/weights_subpops.R
@@ -11,12 +11,12 @@
 #' @examples
 #' # if every individual has a different subpopulation, weights are uniform:
 #' subpops <- 1:10
-#' w <- weightsSubpops(subpops)
+#' w <- weights_subpops(subpops)
 #' stopifnot(all(w == rep.int(1/10,10)))
 #'
 #' # subpopulations can be strings too
 #' subpops <- c('a', 'b', 'c')
-#' w <- weightsSubpops(subpops)
+#' w <- weights_subpops(subpops)
 #' stopifnot(all(w == rep.int(1/3,3)))
 #' 
 #' # if there are two subpopulations
@@ -24,13 +24,19 @@
 #' # then the individuals in this first subpopulation weight half as much 
 #' # as the ones in the second subpopulation
 #' subpops <- c(1, 1, 2)
-#' w <- weightsSubpops(subpops)
+#' w <- weights_subpops(subpops)
 #' stopifnot(all(w == c(1/4,1/4,1/2)))
 #' 
 #' @export
-weightsSubpops <- function(subpops) {
-    if (missing(subpops) || is.null(subpops)) stop('Fatal: subpopulation assignments are missing!')
-    if (anyNA(subpops)) stop('Fatal: subpopulations vector contains NAs!')
+weights_subpops <- function(subpops) {
+    # validate inputs
+    if (missing(subpops))
+        stop('subpopulation assignments are missing!')
+    if (is.null(subpops))
+        stop('subpopulation assignments cannot be NULL!')
+    if (anyNA(subpops))
+        stop('subpopulations vector contains NAs!')
+    
     ## count number of individuals in each subpopulation
     subpop2c <- table(subpops)
     ## count number of subpopulations